줌 이미지 버튼 위치, 이미지 이동 범위 수정
